requirements.txt
aiodns==3.0.0
aiogram==2.22.2
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
Brotli==1.0.9
cchardet==2.1.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
frozenlist==1.3.1
idna==3.4
multidict==6.0.2
psycopg2-binary==2.9.4
pycares==4.2.2
pycparser==2.21
python-dotenv==0.21.0
pytz==2022.4
requests==2.28.1
urllib3==1.26.12
yarl==1.8.1
